Preliminary publishing support, big changes. First of all, having this big diff was a huge mistake. My first two publishing designs were a complete failure, so I couldn't make small commits since I was to break compilation and everything else.
A publish group is a collection of publish group entries, a group entry is
usually a dns service. Groups are necessary to have group entries change as a
whole, for example, if cups publishes both IPP and HTPP services under the same
name (one application, two protocols) it's wise that if you get a conflict in
one service, the other one gets unpublished. One may think that this could be a
job for the application layer, it isn't feasible since the publish notifications
are asynchronous.

Suppose only IPP gets a conflict, you would get an async conflict notification
for it, you would then change the name and retry, but you don't know if HTTP
will conflict also, so you may change HTTP name as well (you want them to have
the same name), and you could later receive a conflict for the first HTTP
name. The application could deal with it by always checking if the notification
is for the last name set, but this is error prone. I'm still not sure about
this, avahi uses this group concept but it really complicates code in mdnsd,
maybe the application should do it, still unsure.

The API should call mdns_group_add(), initiate it's services with
mdns_service_init() and start adding them with mdns_group_add_service(), when
all it's done, these functions populate a group list in mdnsd, which doesn't do
anything until the user issues a mdns_group_commit(), publishing then
starts. All notifications are 'group' oriented, if any entry gets a conflict
user gets a single conflict notification, the group is removed from mdnsd, so
state is *completely* reset, the user should start the whole process again.

All the old publish_* functions were removed as well as the structures, we now
have the following arrangement:

struct pg -> specifies a publish group, it's actually just a container for
publish_entries, has a pointer to controller and some flags. I might kill this
structure and have a relation inside pge, see below.

struct pge -> specifies a publish entry group, or a service, it has a pge_if for
each interface where the entry should be published, pge is a really weak
structure as it's just a holder of pge_if, and pg is almost just a holder of
pge, so I might merge them.

struct pge_if -> specifies a publish entry "interface" for the lack of a better
name, all the publishing intelligence is here, it has a state for each
pge_if_fsm machine, which does the actuall publishing, and holds all the RR
being advertised.

If I merge both pge and pg, design gets much cleaner and flat, the idea of
having 3 nested structures to express one thing bothers me like hell.

Groups are collected in a pg_queue, pge_queue it's just a shortcut to avoid one
TAILQ_FOREACH, when I merge both structures, we should have only one pge_queue.

When pge_if gets published, every RR is linked in an auth_rr_list belonging to
that interface, I'll turn it into a tree one day since O(n) lookups for every
question can be quite expensive.

I hate typing, read the code.