Experimental support for multi-homed setups.
Design is kinda of a hack, but it seems still the best solution, a T_A
resource record now has the option to have addr as INADDR_ANY, if that
is the case, whenever we send that RR out of an interface, the correct
interface address will be patched onto the RR and then restored.
rr_rdata_cmp() now also needs to be smarter to take this into account.

With this solution we have:
	* Conflict resolution is mostly the same.
	* We don't need a rr for every interface, therefore we don't
need a different service for every interface.
	* Can still use the same myname for all interfaces.
	* An uglier code.
	* We have to republish all services whenever an interface comes
up and down, including the addresses (shouldn't be a major issue),
untested.
	* One pge for all primary addresses, instead of one per address.

Also some minor improvements.