Move mdnsl into a static libmdns library
This moves mdnsl around so mdnsctl and mdnsd use the same library.
Due to build infrastructure constraints (using objdirs), there needs
to be an ifdef to use the correct library path to link to.

This diff does not do anything more to establish a library but move
the file around and create libmdns.a.