FIXUP: mcast set/join/leave based on AF
If there are packets going out on multiple interfaces while operating on MDNS
IPV4/6, interfaces should not be expected to be dual-stack. IPv6 only requires
an index, and doesn't seem to have any problems if there are no IPv6 addresses
on the interface. IPv4 requires an address on the interface to do any
multicast operations.