Initial attempt at a reflector.
Allow us to reflect PTR packets between two interfaces/networks in the daemon.
mdnsd gets a -r option with the following syntax.

-r accept@any:any
-r deny@52:54:00:0e:a3:0e:any
-r accept@any:_http._tcp
-r deny@192.168.1.1:_http._tcp
-r accept@192.168.1.1:any

Default policy is to accept if no rules match. Mac address matching is parsed
but not implemented.

We only remove the PTR records from AN and AS section, so we prevent the other
side from browsing such services, this should be the desired functionality in
most setups.